fix upload of board
this fixes three bugs. The first is that there was no report for the
validation errors of the game generation. The second is that there
was no report of validation errors of the input format.

The third was a logic bug where I switched height and width in the
parsing code.